Refactor folder structure
reorganize folder structure with pages and components
