+++ app/routes.ts
@@ -1,3 +1,3 @@
 import { type RouteConfig, index } from "@react-router/dev/routes";
 
-export default [index("routes/home.tsx")] satisfies RouteConfig;
+export default [index("routes/pages/home.tsx")] satisfies RouteConfig;
